--- modules/org-layouts/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "org-layouts",
-  "version": "1.0.1",
-  "private": true,
-  "description": "",
-  "main": "src/index.js",
-  "keywords": [],
-  "author": "",
-  "license": "CC0",
-  "browser": {
-    "querystring": "querystring-browser"
-  },
-  "dependencies": {
-    "debounce": "^1.0.2",
-    "natsort": "^1.0.6",
-    "periodo-ui": "^1.1.0",
-    "periodo-utils": "^5.0.0",
-    "prop-types": "^15.5.10",
-    "querystring-browser": "^1.0.4",
-    "ramda": "^0.24.1",
-    "react": "^16.2.0",
-    "react-geomicons": "^2.1.0",
-    "react-hyperscript": "github:periodo/react-hyperscript",
-    "stream-consume": "^0.1.0",
-    "through2": "^2.0.3"
-  }
-}
