--- modules/org-layouts/src/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  LayoutRenderer: require('./LayoutRenderer'),
-  LayoutEditor: require('./LayoutEditor'),
-
-  blocks: {
-    DOM: require('./DOMBlock'),
-    List: require('./ListBlock'),
-  },
-}
