+++ modules/periodo-app/package.json
@@ -18,9 +18,9 @@
     "jsonld": "^1.0.4",
     "language-tags": "^1.0.5",
     "lov-ns": "^1.1.1",
+    "natsort": "^1.0.6",
     "n3": "^1.0.0-beta.1",
     "org-async-actions": "^1.0.0",
-    "org-layouts": "^1.0.1",
     "org-n3-utils": "^2.0.1",
     "org-shell": "^6.0.0",
     "parse-link-header": "^1.0.1",
